separate url and add dependencis.py
