+++ app/dependencies.py
@@ -1,0 +1,8 @@
+from .database import SessionLocal
+
+async def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
