+++ app/main.py
@@ -1,29 +1,20 @@
-from typing import Annotated
-from datetime import datetime, timedelta
-
 from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.contrib.auth.password_validation import validate_password
-from django.contrib.auth.hashers import (
-    make_password,
-    check_password,
-    is_password_usable,
-)
 
-from fastapi import Query, Depends, FastAPI, HTTPException, APIRouter, status, Request, Header
+from fastapi import FastAPI, Depends, HTTPException, Request
 from fastapi.responses import JSONResponse
 from fastapi.middleware.trustedhost import TrustedHostMiddleware
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.security import HTTPBearer
 from sqlalchemy.orm import Session
 
 import environ
-import jwt
-import requests
-import math
 
-from . import crud, models, schemas
-from .database import SessionLocal, engine
+import jwt
+
+from . import models
+from .database import engine
+from .routers import school, user
+from . import crud, schemas
+from .dependencies import get_db
 
 # 환경변수 init
 env = environ.Env()
@@ -48,33 +39,12 @@ settings.configure(
         },
     ],
 )
-
-
 models.Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
 
-
-# Dependency
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-# Router
-user_router = APIRouter(prefix="/user", tags=["User"])
-school_router = APIRouter(prefix="/school", tags=["School"])
-
-
-# swagger ui 헤더에 jwt 추가를 위한 스키마
-auth_scheme = HTTPBearer()
-
-
 # middleware
-# 토큰 검증 미들웨어
+# 토큰 검증
 @app.middleware("http")
 async def check_access(request: Request, call_next):
     path = request.url.path
@@ -87,7 +57,6 @@ async def check_access(request: Request, call_next):
         "/user/auth/signup",
         "/user/auth/signin",
         "/school/list",
-        "/school/lunch-menu"
     ]
     key = request.headers.get("Authorization")
     print(key)
@@ -101,11 +70,8 @@ async def check_access(request: Request, call_next):
         ]
     else:
         return JSONResponse(status_code=400, content={"message": "토큰 검증에 실패했습니다."})
-
     response = await call_next(request)
     return response
-
-
 # CORS 미들웨어
 app.add_middleware(
     CORSMiddleware,
@@ -115,120 +81,11 @@ app.add_middleware(
     allow_headers=["*"],
 )
 # TrustedHost 미들웨어
-app.add_middleware(TrustedHostMiddleware, allowed_hosts=[
-                   "127.0.0.1", "localhost"])
+app.add_middleware(TrustedHostMiddleware, allowed_hosts=["127.0.0.1", "localhost"])
 
-
-# User
-# 1.회원가입 - 1.이메일 중복검사
-@user_router.post(
-    "/auth/signup/validation",
-    status_code=200,
-    response_model=schemas.Result,
-    responses={
-        400: {
-            "model": schemas.Result,
-            "content": {
-                "application/json": {
-                    "example": {"result": False, "message": "error message"}
-                }
-            },
-        }
-    },
-)
-def is_email_usable(user: schemas.UserBase, db: Session = Depends(get_db)):
-    """
-    `이메일 중복검사`
-    """
-    db_user = crud.get_user(db, email=user.email)
-    if db_user:
-        return JSONResponse(
-            status_code=400, content={"result": False, "message": "이미 가입된 이메일입니다."}
-        )
-        # raise HTTPException(
-        #     status_code=400, detail={"result": False, "message": "이미 가입된 이메일입니다."}
-        # )
-    return {"result": True, "message": "회원가입이 가능한 이메일입니다."}
-
-
-# 1.회원가입 - 2.회원가입
-@user_router.post(
-    "/auth/signup",
-    status_code=200,
-    response_model=schemas.Result,
-    responses={
-        400: {
-            "model": schemas.Result,
-            "content": {
-                "application/json": {
-                    "example": {"result": False, "message": "error message"}
-                }
-            },
-        }
-    },
-)
-def signup(user: schemas.UserCreate, db: Session = Depends(get_db)):
-    """
-    `회원가입`
-    """
-    db_user = crud.get_user(db, email=user.email)
-    if db_user:
-        return JSONResponse(
-            status_code=400, content={"result": False, "message": "이미 가입된 이메일입니다."}
-        )
-
-    # password validation
-    try:
-        validate_password(user.password)
-    except ValidationError as error:
-        message = ""
-        for error_message in error.messages:
-            message = f"{message} {error_message}"
-        raise HTTPException(
-            status_code=422,
-            detail=[
-                {
-                    "loc": ["body", "password"],
-                    "msg": message.lstrip(),
-                    "type": "value_error.password",
-                }
-            ],
-        )
-
-    if crud.create_user(db, user=user):
-        return {"result": True, "message": "회원가입이 완료되었습니다."}
-    else:
-        raise HTTPException(status_code=500)
-
-
-# 2.로그인 - 1.로그인
-@user_router.post("/auth/signin", status_code=200, response_model=schemas.Token)
-def signin(user: schemas.UserSignin, db: Session = Depends(get_db)):
-    """
-    `로그인`
-    """
-    db_user = crud.get_user(db, email=user.email)
-    if db_user is None or not check_password(user.password, db_user.password):
-        raise HTTPException(
-            status_code=400, detail={"result": False, "message": "로그인에 실패하였습니다."}
-        )
-
-    token = jwt.encode(
-        {"email": user.email, "exp": datetime.utcnow() + timedelta(hours=24)},
-        env("JWT_SECRET"),
-        env("JWT_ALGORITHM"),
-    )
-
-    return {"result": True, "message": "로그인 되었습니다.", "token": token}
-
-
-# 3.사용자 정보 조회 - 1.내 정보 조회
-@user_router.get("/info", status_code=200, response_model=schemas.User)
-async def my_info(
-    request: Request, db: Session = Depends(get_db), _: str = Depends(auth_scheme)
-):
-    db_user = crud.get_user(db, email=request.state.email)
-    return db_user
+# router
+app.include_router(user.router)
+app.include_router(school.router)
 
 
 # @app.post("/users/", response_model=schemas.User)
@@ -271,102 +128,3 @@ def read_user(user_id: str, db: Session = Depends(get_db)):
 #     return db_user
 
 
-# school
-niceKEY = env("NICE_API_KEY")
-niceURL = "https://open.neis.go.kr/hub"
-params = {"Type": "json", "KEY": niceKEY}
-# 1. 학교정보 - 1. 학교 정보 검색
-
-
-def filterSchoolList(school):
-    return {
-        "schoolName": school["SCHUL_NM"],
-        "schoolAddress": school["ORG_RDNMA"],
-        "schoolCode": school["SD_SCHUL_CODE"],
-        "areaCode": school["ATPT_OFCDC_SC_CODE"]
-    }
-
-
-@school_router.get("/list", status_code=200, response_model=schemas.SchoolLists)
-# schoolName = None : 모든 학교를 ㄱㄴㄷ 순으로 응답
-async def schoolList(schoolName: str | None = None, pageNumber: int | None = 1, dataSize: int | None = 10):
-    nParams = params.copy()
-    nParams.update(
-        {"SCHUL_NM": schoolName, "pIndex": pageNumber, "pSize": dataSize})
-    response = requests.get(f"{niceURL}/schoolInfo", params=nParams).json()
-    try:
-        response["schoolInfo"]
-    except:
-        code = int(response["RESULT"]["CODE"].split("-")[1])
-        if code == 200 or code == 336:  # 200: 해당하는 데이터가 없습니다. / 336: 데이터요청은 한번에 최대 1,000건을 넘을 수 없습니다
-            raise HTTPException(status_code=400, detail={
-                                "code": 400, "message": response["RESULT"]["MESSAGE"]})
-        else:
-            raise HTTPException(status_code=500, detail={
-                                "code": 500, "message": "내부 API호출 실패"})
-    schoolList = list(map(filterSchoolList, response["schoolInfo"][1]["row"]))
-    totalPageNumber = math.ceil(
-        response["schoolInfo"][0]["head"][0]["list_total_count"]/dataSize)
-    return JSONResponse(status_code=200, content={
-        "schoolList": schoolList,
-        "pagination": {
-            "pageNumber": pageNumber,
-            "dataSize": dataSize,
-            "totalPageNumber": totalPageNumber,
-        }
-    }
-    )
-
-
-# 2. 급식정보 - 1. 급식 정보 조회
-def lunchMenuToDict(str):
-    splited = str.split(" ")
-    return {"menu": splited[0], "allergy": [] if not splited[1] else list(map(int, splited[1].strip("("")").split(".")))}
-
-
-def originToDict(str):
-    splited = str.split(" : ")
-    return {"ingredient": splited[0], "origin": splited[1]}
-
-
-@school_router.get("/lunch-menu", status_code=200, response_model=schemas.SchoolLunch)
-async def schoolLunch(areaCode: str | None, schoolCode: int | None, date: Annotated[int | None, Query(ge=10000000, lt=100000000)]):
-    nParams = params.copy()
-    nParams.update(
-        {"ATPT_OFCDC_SC_CODE": areaCode, "SD_SCHUL_CODE": schoolCode, "MLSV_YMD": date})
-    response = requests.get(
-        f"{niceURL}/mealServiceDietInfo", params=nParams).json()
-    try:
-        response["mealServiceDietInfo"]
-    except:
-        code = int(response["RESULT"]["CODE"].split("-")[1])
-        if code == 200:  # 200: 해당하는 데이터가 없습니다.
-            raise HTTPException(status_code=400, detail={
-                                "code": 400, "message": response["RESULT"]["MESSAGE"]})
-        else:
-            raise HTTPException(status_code=500, detail={
-                                "code": 500, "message": "내부 API호출 실패"})
-    row = response["mealServiceDietInfo"][1]["row"][0]
-    lunchMenu = list(map(lunchMenuToDict, row["DDISH_NM"].split("<br/>")))
-    origin = list(map(originToDict, row["ORPLC_INFO"].split("<br/>")))
-    return JSONResponse(status_code=200, content={
-        "lunchMenu": lunchMenu,
-        "origin": origin
-    })
-
-
-# 뭔지 모르겠지만 school_router로 가져왔습니다.
-
-
-@school_router.get("/{school_code}", response_model=schemas.School)
-def read_school(school_code: str, db: Session = Depends(get_db)):
-    db_school = crud.get_school(db, school_code=school_code)
-    password = "1234"
-    hashed_password = make_password(password)
-    print(check_password(password, hashed_password))
-    print(db_school.users)
-    return db_school
-
-
-app.include_router(user_router)
-app.include_router(school_router)
